+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Possible log types:
 
 - [changed] `SaltyRTC.sendApplicationMessage` now throws an
   `InvalidStateException` if SaltyRTC instance is not in `TASK` state
+- [fixed] Stop logging private key
 
 ### v0.7.1 (2016-11-15)